Refactor claim overview container
