+++ src/portals/hope/common/hooks/use-push-member-history.ts
@@ -1,0 +1,14 @@
+import { useMemberHistory } from 'portals/hope/common/hooks/use-member-history'
+import { useEffect } from 'react'
+
+export const usePushMemberHistory = (memberId?: string | null) => {
+  const { pushToMemberHistory } = useMemberHistory()
+
+  return useEffect(() => {
+    if (!memberId) {
+      return
+    }
+
+    pushToMemberHistory(memberId)
+  }, [memberId])
+}
