+++ src/portals/hope/pages/members/MemberSearchPage.tsx
@@ -12,7 +12,7 @@ import {
 import React, { useRef } from 'react'
 import { useHistory } from 'react-router'
 import { Page } from 'portals/hope/pages/routes'
-import { useMemberHistory } from 'portals/hope/features/user/hooks/use-member-history'
+import { useMemberHistory } from 'portals/hope/common/hooks/use-member-history'
 import { useMemberSearch } from 'portals/hope/features/members-search/hooks/use-member-search'
 import { ExtensiveMemberSearchQuery } from 'types/generated/graphql'
 import { PushUserAction } from 'portals/hope/features/tracking/utils/tags'
